+++ src/app/chapter/[courseName]/card.tsx
@@ -3,7 +3,7 @@ import Link from "next/link";
 export const Card = ({chapter}: {chapter: Chapter}) => {
 
     return (
-        <Link href="/quiz/0" className="flex flex-col justify-around p-2 h-20 border-2 rounded-lg shadow-lg shadow-bg-primary">
+        <Link href="/quiz/1" className="flex flex-col justify-around p-2 h-20 border-2 rounded-lg shadow-lg shadow-bg-primary">
             <div className="font-extrabold">{chapter.name}</div>
             <div className="font-light text-sm">{chapter.context}</div>
         </Link>
